Extracted most controllers from server.js
