+++ errors/error-checks.js
@@ -1,0 +1,2 @@
+exports.DuplicateError = err => err.code === 11000 || err.code === 11001;
+exports.BadRequestError = err => err.name === 'BadRequestError';
