+++ server.js
@@ -9,14 +9,12 @@ const Slide = require('./models/slide.js')
 const mongoose = require('mongoose')
 const bodyParser = require('body-parser')
 const checkAdminRights = require('./helpers').checkAdminRights
+const instagram = require('./controllers/instagram.js')
+const slide = require('./controllers/slide.js')
 
 mongoose.Promise = Promise
 
 let app = express()
-
-const redirect_uri = 
-    settings.OAuth.REDIRECT_URI || 
-    'http://localhost:8888/callback'
 
 // Set up an Express session, which is required for CASAuthentication.
 app.use( session({
@@ -24,8 +22,10 @@ app.use( session({
   resave            : false,
   saveUninitialized : true
 }));
+
 app.use(express.static('public'))
 app.use(bodyParser.json())
+
 // Create a new instance of CASAuthentication.
 var cas = new CASAuthentication({
   cas_url         : settings.cas_url, // The URL of the CAS server.	
@@ -79,99 +79,18 @@ app.get( '/dashboard', cas.bounce, function ( req, res ) { // TODO: USE TEMPLATI
 // Landing portal.
 app.get('/', (req, res) => res.sendFile('./public/index.html'))
 
-app.get('/instagram', cas.block, function(req, res) {
-    access_token = req.query.access_token || null
-    if (access_token){ // todo: check req.originalUrl, check if recently updated and cached.
-        fetch("https://api.instagram.com/v1/users/self/media/recent/?access_token=" + access_token)
-            .then(response => {
-                if (response.status === 200){
-                    // User.cacheToken(access_token)
-                }
-                return response
-            })
-            .then(response => response.json())
-            .then(json => json.data)
-            .then(posts => posts.map(post => {
-                return {
-                    img: post.images.standard_resolution,
-                    caption: post.caption && post.caption.text
-                }
-            }))
-            .then(media => {
-                // Just a temporary example rendering of the images
-                result = ''
-                for (let i=0; i<media.length;i++){
-                    result+='<img src="'+media[i].img.url+'" alt="Oops, unable to find this image"/>'
-                }
-                res.send(result)
-                // res.render('media', media)
-                return media
-            })
-            // .then(media => User.cacheInstagram(app.locals.db, media))
-            .catch(err => {console.log(err);res.sendStatus(500)}
-        )
-    }
-})
-
+app.get('/instagram', cas.block, instagram.getMedia)
 
 // ####################################################################
 //            API for the screen
 // ####################################################################
 
-// Unauthenticated clients will receive a 401 Unauthorized response instead of
-// the JSON data.
-app.get( '/api', cas.block, function ( req, res ) {
-    res.json( { success: true } );
-});
-
-// An example of accessing the CAS user session variable. This could be used to
-// retrieve your own local user records based on authenticated CAS username.
-app.get( '/api/user', cas.block, function ( req, res ) {
-    // At the moment only returns the username of the logged in user.
-    res.json( { cas_user: req.session[ cas.session_name ] } );
-});
-
 // PUBLIC IN ORDER FOR RASPBERRY PI TO ACCESS IT.
-app.get('/api/screen/slides', function(req, res) {
-    // Return all slides
-    Slide.find().then(result => res.json(result))
-});
+app.get('/api/screen/slides', slide.getAllSlides);
 
-app.post('/api/screen/slides/create', cas.block, checkAdminRights(cas), function(req, res) {
-    // Save new slide.
-    if (!req.body) return res.sendStatus(400)
-    Slide.create(req.body)
-        .then(slide => {res.status(201).res.json({"ok":true, slide})})
-        .catch(err => {
-            if (err instanceof mongoose.Error.ValidationError){
-                res.status(400)
-                res.json({'message':'Validation failed',status:400,ok:false,errorName:'ValidationError',errors:err.errors})
-            }
-            else {
-                // Can we catch any other known errors here? Either way we should log it.
-                // This is the downside to not using an error library.
-                // log(err.message, err.errors)
-                res.sendStatus(500)
-            }
-        })
-});
-
-app.get('/api/screen/slides/:id', cas.block, function(req, res) {
-    Slide.find({_id:req.params.id}).then(slide => res.json(slide))
-});
-
-// Currently a POST without a body. Could make the id a required part of the request body
-// and simply make the route [...]/remove.
-app.post('/api/screen/slides/remove/:id', cas.block, checkAdminRights(cas), function(req, res) {
-    // Status is {"ok":1, "n":1} if all went well.
-    if (mongoose.Types.ObjectId.isValid(req.params.id)) {
-        Slide.deleteOne({_id:req.params.id}).then(status => res.json(status))
-    }
-    else { // should this be a 200 with json response {"ok":true, "n":0}?
-        res.status(400)
-        res.json({'message':'Could not find given id', status: 400, ok: false})
-    }
-});
+app.get('/api/screen/slides/:id',         cas.block,                        slide.getById);
+app.post('/api/screen/slides/create',     cas.block, checkAdminRights(cas), slide.create);
+app.post('/api/screen/slides/remove/:id', cas.block, checkAdminRights(cas), slide.removeById);
 
 // ####################################################################
 //            CAS API
@@ -186,51 +105,11 @@ app.get( '/authenticate', cas.bounce_redirect );
 app.get( '/logout', cas.logout );
 
 // ####################################################################
-//            Instagram OAuth, code modified from https://github.com/mpj/oauth-bridge-template
+//            Instagram OAuth
 // ####################################################################
 
-/* Client prompted about app-permissions if not previously approved.
-   Instagram redirects to the redirect_uri (/callback) provided when registering the app,
-   but with a ?code= query parameter.
-*/
-app.get('/instagram/login', cas.block, function(req, res) {
-    res.redirect('https://api.instagram.com/oauth/authorize/?' +
-        querystring.stringify({
-            response_type: 'code',
-            client_id: process.env.INSTAGRAM_CLIENT_ID,
-            redirect_uri,
-        })
-    )
-})
-
-/* Upon receiving the redirect, extract the code from the query and
-   proceed to ask instagram for an access token by providing the code received 
-   and the necessary credentials. After receiving a response from instagram 
-   we redirect the user to a desired frontend destination with the access_token
-   as an extractable query parameter, which is then used as a key for the instagram API.
-*/
-app.get('/callback', function(req, res) {
-    const code = req.query.code || null
-    const authOptions = {
-        url: 'https://api.instagram.com/oauth/access_token',
-        form: {
-            code: code,
-            redirect_uri,
-            grant_type: 'authorization_code',
-            client_id: process.env.INSTAGRAM_CLIENT_ID,
-            client_secret: process.env.INSTAGRAM_CLIENT_SECRET
-        },
-        json: true
-    }
-    request.post(authOptions, function(error, response, body) {
-        if(error){
-            console.log(error)
-        }
-        const access_token = body.access_token
-        const uri = settings.OAuth.FRONTEND_URI || "http://localhost:8888"
-        res.redirect(uri + '?access_token=' + access_token)
-    })
-})
+app.get('/instagram/login', cas.block, instagram.authorize)
+app.get('/instagram/callback', instagram.callback)
 
 // ####################################################################
 //            Launch app to port 8888
